Encryption did not fail if no key for recipient was found.
The problem is rooted in the simplicity of GPGME:Crypto and GPGME:Ctx
encrypt, which both have no problem with missing keys (only with invalid
ones)

Fix verifies that for any pattern in the To: field there exists a key
(found by Ctx.keys).

This is a little clumsy. I would like to
a) have a key selection or search for missing keys
b) have a user verify the choice of encryption keys before sending them
   (unless configured otherwise)